Move Spin and implement SpinResolved.
