+++ gqcp/include/Basis/SpinorBasis/GSpinorBasis.hpp
@@ -22,7 +22,6 @@
 #include "Basis/ScalarBasis/ScalarBasis.hpp"
 #include "Basis/SpinorBasis/RSpinorBasis.hpp"
 #include "Basis/SpinorBasis/SimpleSpinorBasis.hpp"
-#include "Basis/SpinorBasis/Spin.hpp"
 #include "Basis/SpinorBasis/USpinorBasis.hpp"
 #include "Basis/Transformations/JacobiRotationParameters.hpp"
 #include "Molecule/Molecule.hpp"
@@ -30,6 +29,7 @@
 #include "Operator/FirstQuantized/Operator.hpp"
 #include "Operator/SecondQuantized/SQOneElectronOperator.hpp"
 #include "Operator/SecondQuantized/SQTwoElectronOperator.hpp"
+#include "QuantumChemical/Spin.hpp"
 #include "Utilities/aliases.hpp"
 #include "Utilities/literals.hpp"
