+++ gqcp/include/Basis/Transformations/SpinResolvedTransformationMatrix.hpp
@@ -18,8 +18,8 @@
 #pragma once
 
 
-#include "Basis/SpinorBasis/Spin.hpp"
 #include "Basis/Transformations/TransformationMatrix.hpp"
+#include "QuantumChemical/Spin.hpp"
 
 
 namespace GQCP {
