+++ gqcp/include/CMakeLists.txt
@@ -12,4 +12,5 @@ add_subdirectory(Operator)
 add_subdirectory(Processing)
 add_subdirectory(QCMethod)
 add_subdirectory(QCModel)
+add_subdirectory(QuantumChemical)
 add_subdirectory(Utilities)
