+++ gqcp/include/ONVBasis/SpinResolvedFrozenONVBasis.hpp
@@ -18,12 +18,12 @@
 #pragma once
 
 
-#include "Basis/SpinorBasis/Spin.hpp"
 #include "ONVBasis/BaseFrozenCoreONVBasis.hpp"
 #include "ONVBasis/BaseONVBasis.hpp"
 #include "ONVBasis/SpinResolvedONVBasis.hpp"
 #include "ONVBasis/SpinUnresolvedFrozenONVBasis.hpp"
 #include "Operator/SecondQuantized/USQHamiltonian.hpp"
+#include "QuantumChemical/Spin.hpp"
 
 
 namespace GQCP {
