+++ gqcp/include/Operator/SecondQuantized/USQOneElectronOperator.hpp
@@ -18,7 +18,6 @@
 #pragma once
 
 
-#include "Basis/SpinorBasis/Spin.hpp"
 #include "Basis/Transformations/JacobiRotationParameters.hpp"
 #include "Basis/Transformations/TransformationMatrix.hpp"
 #include "DensityMatrix//OneDM.hpp"
@@ -26,6 +25,7 @@
 #include "DensityMatrix//TwoDM.hpp"
 #include "Mathematical/Functions/ScalarFunction.hpp"
 #include "Mathematical/Representation/QCMatrix.hpp"
+#include "QuantumChemical/Spin.hpp"
 #include "Utilities/type_traits.hpp"
 
 #include <array>
