+++ gqcp/include/QCModel/HF/UHF.hpp
@@ -18,13 +18,13 @@
 #pragma once
 
 
-#include "Basis/SpinorBasis/Spin.hpp"
 #include "Basis/Transformations/TransformationMatrix.hpp"
 #include "DensityMatrix//OneDM.hpp"
 #include "DensityMatrix//SpinResolvedOneDM.hpp"
 #include "Mathematical/Representation/Matrix.hpp"
 #include "Operator/SecondQuantized/USQOneElectronOperator.hpp"
 #include "QCModel/HF/RHF.hpp"
+#include "QuantumChemical/Spin.hpp"
 
 
 namespace GQCP {
