+++ gqcp/include/QuantumChemical/CMakeLists.txt
@@ -1,0 +1,5 @@
+target_sources(gqcp
+    PRIVATE
+        Spin.hpp
+        SpinResolved.hpp
+)
