+++ gqcp/include/QuantumChemical/Spin.hpp
