+++ gqcp/include/gqcp.hpp
@@ -52,7 +52,6 @@
 #include "Basis/SpinorBasis/OrbitalSpace.hpp"
 #include "Basis/SpinorBasis/RSpinorBasis.hpp"
 #include "Basis/SpinorBasis/SimpleSpinorBasis.hpp"
-#include "Basis/SpinorBasis/Spin.hpp"
 #include "Basis/SpinorBasis/USpinorBasis.hpp"
 #include "Basis/Transformations/JacobiRotationParameters.hpp"
 #include "Basis/Transformations/OrbitalRotationGenerators.hpp"
@@ -110,12 +109,12 @@
 #include "Mathematical/Optimization/NonLinearEquation/NonLinearEquationSolver.hpp"
 #include "Mathematical/Optimization/NonLinearEquation/step.hpp"
 #include "Mathematical/Optimization/OptimizationEnvironment.hpp"
+#include "Mathematical/Representation/#include " Mathematical / Representation / QCRankFourTensor.hpp "
 #include "Mathematical/Representation/DenseVectorizer.hpp"
 #include "Mathematical/Representation/ImplicitMatrixSlice.hpp"
 #include "Mathematical/Representation/ImplicitRankFourTensorSlice.hpp"
 #include "Mathematical/Representation/Matrix.hpp"
 #include "Mathematical/Representation/MatrixRepresentationEvaluationContainer.hpp"
-#include "Mathematical/Representation/#include "Mathematical/Representation/QCRankFourTensor.hpp"
 #include "Mathematical/Representation/SquareMatrix.hpp"
 #include "Mathematical/Representation/SquareRankFourTensor.hpp"
 #include "Mathematical/Representation/StorageArray.hpp"
@@ -238,6 +237,7 @@
 #include "QCModel/HF/GHF.hpp"
 #include "QCModel/HF/RHF.hpp"
 #include "QCModel/HF/UHF.hpp"
+#include "QuantumChemical/Spin.hpp"
 #include "Utilities/CRTP.hpp"
 #include "Utilities/aliases.hpp"
 #include "Utilities/linalg.hpp"
