+++ gqcpy/src/QuantumChemical/CMakeLists.txt
@@ -1,0 +1,5 @@
+list(APPEND python_bindings_sources
+    ${CMAKE_CURRENT_SOURCE_DIR}/Spin_bindings.cpp
+)
+
+set(python_bindings_sources ${python_bindings_sources} PARENT_SCOPE)
