+++ gqcpy/src/QuantumChemical/Spin_bindings.cpp
@@ -15,7 +15,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with GQCG-GQCP.  If not, see <http://www.gnu.org/licenses/>.
 
-#include "Basis/SpinorBasis/Spin.hpp"
+#include "QuantumChemical/Spin.hpp"
 
 #include <pybind11/pybind11.h>
